requirements.txt
fastapi
uvicorn
jinja2
httpx
python-dotenv
rich==13.7.1
requests
beautifulsoup4
google-api-python-client
google-auth
google-auth-oauthlib
itsdangerous